requirements.txt
python-docx==0.8.*
urlextract==1.6.*
